requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
av==17.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.7
ctranslate2==4.7.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.1
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fsspec==2026.2.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
groq==1.1.2
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.23.2
packaging==26.0
pillow==11.3.0
proglog==0.1.12
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pyreadline3==3.5.4
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
sentencepiece==0.2.1
shellingham==1.5.4
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.19.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==4.44.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.7
youtube-transcript-api==1.2.4
yt-dlp==2026.3.17
